py/requirements.txt
matplotlib==3.0.0
numpy==1.17.0
pandas==0.23.4
Pillow==6.1.0
scipy==1.2.1
torch==1.2.0
torchvision==0.4.0

